requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8
djangorestframework==3.12.1
django-webpack-loader>=1.0.0
drf-yasg==1.20.0